refactor: set up deployer-lib; init app-deployer basic structure
