+++ deployer/app-deployer/go.mod
@@ -1,0 +1,8 @@
+module github.com/RHEcosystemAppEng/SaaSi/deployer/app-deployer
+
+go 1.19
+
+require (
+	github.com/gorilla/mux v1.8.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
